requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.5.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.2.0
pydantic==2.5.0
pydantic-settings==2.5.0
reportlab==4.0.7
python-dotenv==1.0.0
pytz==2023.3
python-dateutil==2.8.2
aiohttp==3.9.1
python-socketio==5.11.0